Place hooks in a directory instead of all in one file
