+++ crates/vm/levm/src/hooks/hook.rs
@@ -1,0 +1,20 @@
+use crate::{
+    call_frame::CallFrame,
+    errors::{ExecutionReport, VMError},
+    vm::VM,
+};
+
+pub trait Hook {
+    fn prepare_execution(
+        &self,
+        vm: &mut VM,
+        initial_call_frame: &mut CallFrame,
+    ) -> Result<(), VMError>;
+
+    fn finalize_execution(
+        &self,
+        vm: &mut VM,
+        initial_call_frame: &CallFrame,
+        report: &mut ExecutionReport,
+    ) -> Result<(), VMError>;
+}
